[feat] split simpson dataset and parsing mech
